collect-infor-py/requirements.txt
alembic==1.8.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
async-timeout==4.0.2
blinker==1.5
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
flash==1.0.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
prettytable==3.4.1
pycparser==2.21
pyecharts==1.9.1
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
redis==4.3.4
simplejson==3.17.6
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.4.44
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.10.0
